+++ apps/extension/src/components/index.tsx
@@ -1,1 +1,2 @@
+export * from "@components/Footer";
 export * from "@components/Header";