+++ apps/extension/src/configurations/index.ts
@@ -1,0 +1,2 @@
+export * from "./theme";
+export * from "./i18n";