+++ apps/extension/src/constants/dateFormats/index.ts
@@ -1,0 +1,32 @@
+export const DateFormats = [
+  {
+    label: "DD/MM/YYYY",
+    example: "04/07/2025",
+    formatString: "dd/MM/yyyy",
+  },
+  {
+    label: "YYYY-MM-DD",
+    example: "2025-07-04",
+    formatString: "yyyy-MM-dd",
+  },
+  {
+    label: "MM/DD/YYYY",
+    example: "07/04/2025",
+    formatString: "MM/dd/yyyy",
+  },
+  {
+    label: "DD.MM.YYYY",
+    example: "04.07.2025",
+    formatString: "dd.MM.yyyy",
+  },
+  {
+    label: "D MMM YYYY",
+    example: "4 Jul 2025",
+    formatString: "d MMM yyyy",
+  },
+  {
+    label: "dddd, D MMMM YYYY",
+    example: "Friday, 4 July 2025",
+    formatString: "EEEE, d MMMM yyyy",
+  },
+];