+++ apps/extension/src/constants/languages/index.ts
@@ -1,15 +1,17 @@
-export enum Languages {
-  English = "en",
-  Romanian = "ro",
-  German = "de",
-  French = "fr",
-  Spanish = "es",
-  Italian = "it",
-  Dutch = "nl",
-  Polish = "pl",
-  Portuguese = "pt",
-  Russian = "ru",
-  Chinese = "zh",
-  Japanese = "ja",
-  Korean = "ko",
-}
+import type { LanguageModel } from "@models";
+
+export const Languages: LanguageModel[] = [
+  { name: "English", value: "en", flag: "🇬🇧" },
+  { name: "Română", value: "ro", flag: "🇷🇴" },
+  { name: "Deutsch", value: "de", flag: "🇩🇪" },
+  { name: "Français", value: "fr", flag: "🇫🇷" },
+  { name: "Español", value: "es", flag: "🇪🇸" },
+  { name: "Italiano", value: "it", flag: "🇮🇹" },
+  { name: "Nederlands", value: "nl", flag: "🇳🇱" },
+  { name: "Polski", value: "pl", flag: "🇵🇱" },
+  { name: "Português", value: "pt", flag: "🇵🇹" },
+  { name: "Русский", value: "ru", flag: "🇷🇺" },
+  { name: "中文", value: "zh", flag: "🇨🇳" },
+  { name: "日本語", value: "ja", flag: "🇯🇵" },
+  { name: "한국어", value: "ko", flag: "🇰🇷" },
+];