+++ apps/extension/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { ReduxProvider, StyleProvider } from "@providers";
+import { I18nProvider, ReduxProvider, StyleProvider } from "@providers";
 import { RouterProvider } from "react-router";
 import { routes } from "./routes";
 
@@ -8,7 +8,9 @@ createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <ReduxProvider>
       <StyleProvider>
-        <RouterProvider router={routes} />
+        <I18nProvider>
+          <RouterProvider router={routes} />
+        </I18nProvider>
       </StyleProvider>
     </ReduxProvider>
   </StrictMode>