+++ apps/extension/src/models/ApplicationSettings.model.ts
@@ -1,12 +1,8 @@
-import type { Languages } from "@constants";
-
-export type Language = keyof typeof Languages;
 export type Theme = "light" | "dark";
 
 export interface ApplicationSettingsModel {
   theme: Theme;
   useSystemTheme: boolean;
-  language: Language;
   showIPV6: boolean;
   showPublicIPNotification: boolean;
   showIconInToolbar: boolean;