+++ apps/extension/src/models/Language.model.ts
@@ -1,0 +1,5 @@
+export interface LanguageModel {
+  name: string;
+  value: string;
+  flag: string;
+}