+++ apps/extension/src/models/index.ts
@@ -1,1 +1,2 @@
 export * from "@models/ApplicationSettings.model";
+export * from "@models/Language.model";