+++ apps/extension/src/pages/Settings/index.tsx
@@ -11,22 +11,39 @@ import {
   Portal,
   RadioGroup,
   Select,
+  Span,
+  Stack,
   Switch,
 } from "@chakra-ui/react";
 import { FaArrowLeft } from "react-icons/fa";
-import { Routes, Languages } from "@constants";
+import { Routes, Languages, DateFormats, TimeFormats } from "@constants";
 import { Header } from "@components";
 import { HiCheck, HiX } from "react-icons/hi";
 
 import { FiClock, FiSettings, FiWifi } from "react-icons/fi";
+import { useTranslation } from "react-i18next";
 
 export const SettingsPage: FC = () => {
   const navigate = useNavigate();
+  const { t, i18n } = useTranslation();
 
   const languages = createListCollection({
-    items: Object.values(Languages),
-    itemToString: (language) => language,
-    itemToValue: (language) => language,
+    items: Languages,
+    itemToString: (language) =>
+      `${language.flag} ${t(`languages.${language.value}`)}`,
+    itemToValue: (language) => language.value,
+  });
+
+  const dateFormats = createListCollection({
+    items: DateFormats,
+    itemToString: (dateFormat) => dateFormat.label,
+    itemToValue: (dateFormat) => dateFormat.formatString,
+  });
+
+  const timeFormats = createListCollection({
+    items: TimeFormats,
+    itemToString: (timeFormat) => timeFormat.label,
+    itemToValue: (timeFormat) => timeFormat.formatString,
   });
 
   const handleNavigateToHomepage = () => {
@@ -44,17 +61,17 @@ export const SettingsPage: FC = () => {
         >
           <FaArrowLeft />
         </IconButton>
-        <Heading size="2xl">Settings</Heading>
+        <Heading size="2xl">{t("settings")}</Heading>
       </Header>
       <Box display="flex" flexDirection="column" gap="6" paddingY="5">
         <Box display="flex" flexDirection="column" gap="3" fontSize="lg">
           <HStack display="flex">
             <Icon as={FiSettings} />
-            <Heading>General</Heading>
+            <Heading>{t("general")}</Heading>
           </HStack>
 
           <Fieldset.Root>
-            <Fieldset.Legend fontSize="lg">Theme</Fieldset.Legend>
+            <Fieldset.Legend fontSize="lg">{t("theme")}</Fieldset.Legend>
             <RadioGroup.Root size="lg">
               <HStack
                 display="flex"
@@ -64,29 +81,39 @@ export const SettingsPage: FC = () => {
                 <RadioGroup.Item value="asd">
                   <RadioGroup.ItemHiddenInput />
                   <RadioGroup.ItemIndicator />
-                  <RadioGroup.ItemText fontSize="lg">Dark</RadioGroup.ItemText>
-                </RadioGroup.Item>
-                <RadioGroup.Item value="asd">
-                  <RadioGroup.ItemHiddenInput />
-                  <RadioGroup.ItemIndicator />
-                  <RadioGroup.ItemText fontSize="lg">Dark</RadioGroup.ItemText>
+                  <RadioGroup.ItemText fontSize="lg">
+                    {t("dark")}
+                  </RadioGroup.ItemText>
                 </RadioGroup.Item>
                 <RadioGroup.Item value="asd">
                   <RadioGroup.ItemHiddenInput />
                   <RadioGroup.ItemIndicator />
                   <RadioGroup.ItemText fontSize="lg">
-                    System
+                    {t("light")}
+                  </RadioGroup.ItemText>
+                </RadioGroup.Item>
+                <RadioGroup.Item value="asd">
+                  <RadioGroup.ItemHiddenInput />
+                  <RadioGroup.ItemIndicator />
+                  <RadioGroup.ItemText fontSize="lg">
+                    {t("system")}
                   </RadioGroup.ItemText>
                 </RadioGroup.Item>
               </HStack>
             </RadioGroup.Root>
           </Fieldset.Root>
-          <Select.Root size="lg" collection={languages}>
+          <Select.Root
+            size="lg"
+            collection={languages}
+            multiple={false}
+            defaultValue={[i18n.language]}
+            onValueChange={(da) => i18n.changeLanguage(da.value[0])}
+          >
             <Select.HiddenSelect />
-            <Select.Label fontSize="lg">Languages</Select.Label>
+            <Select.Label fontSize="lg">{t("language")}</Select.Label>
             <Select.Control>
               <Select.Trigger>
-                <Select.ValueText placeholder="Select language" />
+                <Select.ValueText placeholder={t("selectLanguage")} />
               </Select.Trigger>
               <Select.IndicatorGroup>
                 <Select.Indicator />
@@ -95,9 +122,10 @@ export const SettingsPage: FC = () => {
             <Portal>
               <Select.Positioner>
                 <Select.Content>
-                  {languages.items.map((framework) => (
-                    <Select.Item item={framework} key={framework}>
-                      {framework}
+                  {languages.items.map((language) => (
+                    <Select.Item item={language.value} key={language.value}>
+                      {language.flag}
+                      {t(`languages.${language.value}`)}
                       <Select.ItemIndicator />
                     </Select.Item>
                   ))}
@@ -105,12 +133,12 @@ export const SettingsPage: FC = () => {
               </Select.Positioner>
             </Portal>
           </Select.Root>
-          <Select.Root size="lg" collection={languages}>
+          <Select.Root size="lg" collection={dateFormats}>
             <Select.HiddenSelect />
-            <Select.Label fontSize="lg">Date/Time format</Select.Label>
+            <Select.Label fontSize="lg">{t("dateFormat")}</Select.Label>
             <Select.Control>
               <Select.Trigger>
-                <Select.ValueText placeholder="Select date/time format" />
+                <Select.ValueText placeholder={t("selectDateFormat")} />
               </Select.Trigger>
               <Select.IndicatorGroup>
                 <Select.Indicator />
@@ -119,9 +147,49 @@ export const SettingsPage: FC = () => {
             <Portal>
               <Select.Positioner>
                 <Select.Content>
-                  {languages.items.map((framework) => (
-                    <Select.Item item={framework} key={framework}>
-                      {framework}
+                  {dateFormats.items.map((dateFormat) => (
+                    <Select.Item
+                      item={dateFormat.formatString}
+                      key={dateFormat.formatString}
+                    >
+                      <Stack gap="0">
+                        <Select.ItemText>{dateFormat.label}</Select.ItemText>
+                        <Span color="fg.muted" textStyle="xs">
+                          {dateFormat.example}
+                        </Span>
+                      </Stack>
+                      <Select.ItemIndicator />
+                    </Select.Item>
+                  ))}
+                </Select.Content>
+              </Select.Positioner>
+            </Portal>
+          </Select.Root>
+          <Select.Root size="lg" collection={timeFormats}>
+            <Select.HiddenSelect />
+            <Select.Label fontSize="lg">{t("timeFormat")}</Select.Label>
+            <Select.Control>
+              <Select.Trigger>
+                <Select.ValueText placeholder={t("selectTimeFormat")} />
+              </Select.Trigger>
+              <Select.IndicatorGroup>
+                <Select.Indicator />
+              </Select.IndicatorGroup>
+            </Select.Control>
+            <Portal>
+              <Select.Positioner>
+                <Select.Content>
+                  {timeFormats.items.map((timeFormat) => (
+                    <Select.Item
+                      item={timeFormat.formatString}
+                      key={timeFormat.formatString}
+                    >
+                      <Stack gap="0">
+                        <Select.ItemText>{timeFormat.label}</Select.ItemText>
+                        <Span color="fg.muted" textStyle="xs">
+                          {timeFormat.example}
+                        </Span>
+                      </Stack>
                       <Select.ItemIndicator />
                     </Select.Item>
                   ))}
@@ -131,7 +199,7 @@ export const SettingsPage: FC = () => {
           </Select.Root>
           <Switch.Root size="lg" display="flex" justifyContent="space-between">
             <Switch.Label fontSize="lg">
-              Show public IP notifications
+              {t("showPublicIpNotifications")}
             </Switch.Label>
             <Switch.HiddenInput />
             <Switch.Control>
@@ -143,7 +211,7 @@ export const SettingsPage: FC = () => {
             </Switch.Control>
           </Switch.Root>
           <Switch.Root size="lg" display="flex" justifyContent="space-between">
-            <Switch.Label fontSize="lg">Show icon in toolbar</Switch.Label>
+            <Switch.Label fontSize="lg">{t("showIconInToolbar")}</Switch.Label>
             <Switch.HiddenInput />
             <Switch.Control>
               <Switch.Thumb>
@@ -157,11 +225,11 @@ export const SettingsPage: FC = () => {
         <Box display="flex" flexDirection="column" gap="3" fontSize="lg">
           <HStack display="flex">
             <Icon as={FiWifi} />
-            <Heading>IP Management</Heading>
+            <Heading>{t("ipManagement")}</Heading>
           </HStack>
 
           <Switch.Root size="lg" display="flex" justifyContent="space-between">
-            <Switch.Label fontSize="lg">Enable IPV6</Switch.Label>
+            <Switch.Label fontSize="lg">{t("enableIPV6")}</Switch.Label>
             <Switch.HiddenInput />
             <Switch.Control>
               <Switch.Thumb>
@@ -175,15 +243,15 @@ export const SettingsPage: FC = () => {
         <Box display="flex" flexDirection="column" gap="3" fontSize="lg">
           <HStack display="flex">
             <Icon as={FiClock} />
-            <Heading>History</Heading>
+            <Heading>{t("history")}</Heading>
           </HStack>
 
           <Select.Root size="lg" collection={languages}>
             <Select.HiddenSelect />
-            <Select.Label fontSize="lg">Number of IPs to show</Select.Label>
+            <Select.Label fontSize="lg">{t("numbersIPsToShow")}</Select.Label>
             <Select.Control>
               <Select.Trigger>
-                <Select.ValueText placeholder="Select number of IPs to show" />
+                <Select.ValueText placeholder={t("selectNumbersOfIPsToShow")} />
               </Select.Trigger>
               <Select.IndicatorGroup>
                 <Select.Indicator />
@@ -192,9 +260,9 @@ export const SettingsPage: FC = () => {
             <Portal>
               <Select.Positioner>
                 <Select.Content>
-                  {languages.items.map((framework) => (
-                    <Select.Item item={framework} key={framework}>
-                      {framework}
+                  {languages.items.map((language) => (
+                    <Select.Item item={language.value} key={language.value}>
+                      {language.name}
                       <Select.ItemIndicator />
                     </Select.Item>
                   ))}