+++ apps/extension/src/providers/index.tsx
@@ -1,2 +1,3 @@
 export * from "@providers/Redux";
 export * from "@providers/Style";
+export * from "@providers/I18n";